# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitPulse Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterRequestFitnessGoal =
  (typeof RegisterRequestFitnessGoal)[keyof typeof RegisterRequestFitnessGoal];

export const RegisterRequestFitnessGoal = {
  weight_loss: "weight_loss",
  muscle_gain: "muscle_gain",
  endurance: "endurance",
  flexibility: "flexibility",
  general_fitness: "general_fitness",
} as const;

export type RegisterRequestActivityLevel =
  (typeof RegisterRequestActivityLevel)[keyof typeof RegisterRequestActivityLevel];

export const RegisterRequestActivityLevel = {
  sedentary: "sedentary",
  lightly_active: "lightly_active",
  moderately_active: "moderately_active",
  very_active: "very_active",
  extra_active: "extra_active",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  fitnessGoal: RegisterRequestFitnessGoal;
  age?: number;
  weight?: number;
  height?: number;
  activityLevel?: RegisterRequestActivityLevel;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  fitnessGoal: string;
  age?: number;
  weight?: number;
  height?: number;
  activityLevel?: string;
  avatarUrl?: string;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  fitnessGoal?: string;
  age?: number;
  weight?: number;
  height?: number;
  activityLevel?: string;
  avatarUrl?: string;
  bio?: string;
  createdAt?: string;
}

export interface UpdateProfileRequest {
  userId: number;
  name?: string;
  age?: number;
  weight?: number;
  height?: number;
  activityLevel?: string;
  fitnessGoal?: string;
  bio?: string;
}

export interface UserStats {
  totalWorkouts: number;
  totalCaloriesBurned: number;
  totalMinutes: number;
  currentStreak: number;
  longestStreak?: number;
  thisWeekWorkouts?: number;
  thisWeekCalories?: number;
  bmi?: number;
  bmiCategory?: string;
}

export interface WorkoutExercise {
  exerciseId?: number;
  exerciseName: string;
  sets?: number;
  reps?: number;
  weight?: number;
  duration?: number;
}

export interface Workout {
  id: number;
  userId: number;
  workoutName: string;
  category?: string;
  duration: number;
  caloriesBurned: number;
  exercises?: WorkoutExercise[];
  notes?: string;
  completedAt?: string;
  createdAt?: string;
}

export interface CreateWorkoutRequest {
  userId: number;
  workoutName: string;
  category?: string;
  duration: number;
  caloriesBurned: number;
  exercises?: WorkoutExercise[];
  notes?: string;
  completedAt?: string;
}

export interface UpdateWorkoutRequest {
  workoutName?: string;
  duration?: number;
  caloriesBurned?: number;
  notes?: string;
}

export type WorkoutPlanLevel =
  (typeof WorkoutPlanLevel)[keyof typeof WorkoutPlanLevel];

export const WorkoutPlanLevel = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface WorkoutPlan {
  id: number;
  name: string;
  description?: string;
  level: WorkoutPlanLevel;
  duration: number;
  category: string;
  isPremium?: boolean;
  exercises?: string[];
}

export type ExerciseDifficulty =
  (typeof ExerciseDifficulty)[keyof typeof ExerciseDifficulty];

export const ExerciseDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Exercise {
  id: number;
  name: string;
  category: string;
  muscleGroup: string;
  difficulty: ExerciseDifficulty;
  description?: string;
  instructions?: string[];
  imageUrl?: string;
  videoUrl?: string;
  caloriesPerMinute?: number;
  equipment?: string;
}

export interface ProgressEntry {
  id: number;
  userId: number;
  date: string;
  workoutsCompleted: number;
  caloriesBurned?: number;
  minutesActive?: number;
  weight?: number;
  steps?: number;
  mood?: string;
}

export interface LogProgressRequest {
  userId: number;
  date: string;
  workoutsCompleted: number;
  caloriesBurned?: number;
  minutesActive?: number;
  weight?: number;
  steps?: number;
  mood?: string;
}

export type ProgressSummaryWeeklyCaloriesItem = {
  day: string;
  calories: number;
};

export type ProgressSummaryMonthlyWorkoutsItem = {
  week: string;
  count: number;
};

export interface ProgressSummary {
  weeklyCalories: ProgressSummaryWeeklyCaloriesItem[];
  monthlyWorkouts: ProgressSummaryMonthlyWorkoutsItem[];
  totalThisMonth: number;
  averageCaloriesPerWorkout?: number;
}

export interface Goal {
  id: number;
  userId: number;
  title: string;
  description?: string;
  targetValue: number;
  currentValue: number;
  unit?: string;
  deadline?: string;
  completed?: boolean;
  category?: string;
}

export interface CreateGoalRequest {
  userId: number;
  title: string;
  description?: string;
  targetValue: number;
  currentValue?: number;
  unit?: string;
  deadline?: string;
  category?: string;
}

export interface UpdateGoalRequest {
  currentValue?: number;
  completed?: boolean;
  title?: string;
  targetValue?: number;
}

export interface Achievement {
  id: number;
  userId: number;
  title: string;
  description: string;
  icon: string;
  unlockedAt?: string;
  category?: string;
}

export type ForgotPasswordBody = {
  email: string;
};

export type GetUserProfileParams = {
  userId: number;
};

export type GetUserStatsParams = {
  userId: number;
};

export type GetWorkoutsParams = {
  userId: number;
  limit?: number;
};

export type GetWorkoutPlansParams = {
  userId?: number;
};

export type GetExercisesParams = {
  category?: string;
  search?: string;
};

export type GetProgressParams = {
  userId: number;
  days?: number;
};

export type GetProgressSummaryParams = {
  userId: number;
};

export type GetGoalsParams = {
  userId: number;
};

export type GetAchievementsParams = {
  userId: number;
};