# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FitPulse Fitness Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user fitness stats
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
...
  /progress:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get user progress data
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Progress data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressEntry"
    post:
      operationId: logProgress
      tags: [progress]
      summary: Log progress entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogProgressRequest"
      responses:
        "201":
          description: Progress logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressEntry"
  /progress/summary:
    get:
      operationId: getProgressSummary
      tags: [progress]
      summary: Get progress summary
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressSummary"
  /goals:
    get:
      operationId: getGoals
      tags: [users]
      summary: Get user goals
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [users]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalRequest"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    put:
      operationId: updateGoal
      tags: [users]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalRequest"
      responses:
        "200":
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fitnessGoal:
          type: string
          enum: [weight_loss, muscle_gain, endurance, flexibility, general_fitness]
        age:
          type: integer
        weight:
          type: number
        height:
          type: number
        activityLevel:
          type: string
          enum: [sedentary, lightly_active, moderately_active, very_active, extra_active]
      required: [name, email, password, fitnessGoal]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        fitnessGoal:
          type: string
        age:
          type: integer
        weight:
          type: number
        height:
          type: number
        activityLevel:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, fitnessGoal]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        fitnessGoal:
          type: string
        age:
          type: integer
        weight:
          type: number
        height:
          type: number
        activityLevel:
          type: string
        avatarUrl:
          type: string
        bio:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email]
    UpdateProfileRequest:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        age:
          type: integer
        weight:
          type: number
        height:
          type: number
        activityLevel:
          type: string
        fitnessGoal:
          type: string
        bio:
          type: string
      required: [userId]
    UserStats:
      type: object
      properties:
        totalWorkouts:
          type: integer
        totalCaloriesBurned:
          type: integer
        totalMinutes:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        thisWeekWorkouts:
          type: integer
        thisWeekCalories:
          type: integer
        bmi:
          type: number
        bmiCategory:
          type: string
      required: [totalWorkouts, totalCaloriesBurned, totalMinutes, currentStreak]
...
    ProgressEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        date:
          type: string
          format: date
        workoutsCompleted:
          type: integer
        caloriesBurned:
          type: integer
        minutesActive:
          type: integer
        weight:
          type: number
        steps:
          type: integer
        mood:
          type: string
      required: [id, userId, date, workoutsCompleted]
    LogProgressRequest:
      type: object
      properties:
        userId:
          type: integer
        date:
          type: string
          format: date
        workoutsCompleted:
          type: integer
        caloriesBurned:
          type: integer
        minutesActive:
          type: integer
        weight:
          type: number
        steps:
          type: integer
        mood:
          type: string
      required: [userId, date, workoutsCompleted]
    ProgressSummary:
      type: object
      properties:
        weeklyCalories:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              calories:
                type: integer
            required: [day, calories]
        monthlyWorkouts:
          type: array
          items:
            type: object
            properties:
              week:
                type: string
              count:
                type: integer
            required: [week, count]
        totalThisMonth:
          type: integer
        averageCaloriesPerWorkout:
          type: integer
      required: [weeklyCalories, monthlyWorkouts, totalThisMonth]
    Goal:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        targetValue:
          type: number
        currentValue:
          type: number
        unit:
          type: string
        deadline:
          type: string
          format: date
        completed:
          type: boolean
        category:
          type: string
      required: [id, userId, title, targetValue, currentValue]
    CreateGoalRequest:
      type: object
      properties:
        userId:
          type: integer
        title:
          type: string
        description:
          type: string
        targetValue:
          type: number
        currentValue:
          type: number
        unit:
          type: string
        deadline:
          type: string
          format: date
        category:
          type: string
      required: [userId, title, targetValue]
    UpdateGoalRequest:
      type: object
      properties:
        currentValue:
          type: number
        completed:
          type: boolean
        title:
          type: string
        targetValue:
          type: number
...