# lib/api-zod/src/generated/types/logProgressRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FitPulse Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface LogProgressRequest {
  userId: number;
  date: Date;
  workoutsCompleted: number;
  caloriesBurned?: number;
  minutesActive?: number;
  weight?: number;
  steps?: number;
  mood?: string;
}